+++ game-warhammer40kroguetrader/CHANGELOG.md
@@ -4,6 +4,12 @@
 
 - None Planned
 
+## [0.1.1] - 2025-01-16
+
+- Fixed: Owlcat mod installer removing "Bundles" folder and files as they have no file extension
+- Added: Button to open the "OwlcatModificationManagerSettings.json" file
+- Added: Notification for Xbox users that custom Portaits mods will not be loaded by that version of the game
+
 ## [0.1.0] - 2025-01-13
 
 - Inital Release